Refactoring: split project into multiple sources
Split project code into multiple files:
- errors.rs - Templaar-specific errors
- new.rs - Handling of the 'new' command
- take.rs - Handling of the 'take' command
- utils.rs - General-purpose utilities

Also add documentation comments where useful.
